tree hash

+++ src/util.rs
@@ -1,5 +1,5 @@
 /// log2 of x floored
-pub(crate) fn log2(x: u128) -> u128 {
+pub(crate) fn log2(x: u32) -> u32 {
     if x == 0 {
         return 0;
     }
@@ -12,7 +12,7 @@ pub(crate) fn log2(x: u128) -> u128 {
 }
 
 /// The number of nodes needed to represent a tree with n leaves.
-pub(crate) fn num_nodes(x: u128) -> u128 {
+pub(crate) fn num_nodes(x: u32) -> u32 {
     if x == 0 {
         0
     } else {
@@ -20,7 +20,7 @@ pub(crate) fn num_nodes(x: u128) -> u128 {
     }
 }
 
-pub(crate) fn level(x: u128) -> u128 {
+pub(crate) fn level(x: u32) -> u32 {
     // XXX: masked log2 really
     if x & 0x01 == 0 {
         return 0;
@@ -33,7 +33,7 @@ pub(crate) fn level(x: u128) -> u128 {
     k
 }
 
-pub(crate) fn left(x: u128) -> u128 {
+pub(crate) fn left(x: u32) -> u32 {
     let k = level(x);
     if k == 0 {
         panic!("leaf nodes have no children");
@@ -42,7 +42,7 @@ pub(crate) fn left(x: u128) -> u128 {
     x ^ (0x01 << (k - 1))
 }
 
-pub(crate) fn right(x: u128) -> u128 {
+pub(crate) fn right(x: u32) -> u32 {
     let k = level(x);
     if k == 0 {
         panic!("leaf nodes have no children");
@@ -54,7 +54,7 @@ pub(crate) fn right(x: u128) -> u128 {
     r
 }
 
-pub(crate) fn sibling(x: u128) -> u128 {
+pub(crate) fn sibling(x: u32) -> u32 {
     let p = parent(x);
     if x < p {
         right(p)
@@ -63,14 +63,14 @@ pub(crate) fn sibling(x: u128) -> u128 {
     }
 }
 
-pub(crate) fn parent(x: u128) -> u128 {
+pub(crate) fn parent(x: u32) -> u32 {
     let k = level(x);
     // println!(" >>> x: {}\nk: {}", x, k+1);
     let b = (x >> (k + 1)) & 0x01;
     (x | (1 << k)) ^ (b << (k + 1))
 }
 
-pub(crate) fn direct_path(x: u128, r: u128) -> Vec<u128> {
+pub(crate) fn direct_path(x: u32, r: u32) -> Vec<u32> {
     let mut d = Vec::new();
     if x == r {
         return d;
